+++ src/gospatial/app/generalHandlers.go
@@ -1,6 +1,7 @@
 package app
 
 import (
+	"fmt"
 	"github.com/gorilla/mux"
 	"html/template"
 	"net/http"
@@ -18,6 +19,10 @@ func MapHandler(w http.ResponseWriter, r *http.Request) {
 		tmpl, _ := template.ParseFiles(map_tmpl)
 		Info.Println(r.RemoteAddr, "GET /map/"+ds+" [200]")
 		tmpl.Execute(w, MapData{Datasource: ds})
+	} else if r.FormValue("apikey") != "" {
+		Warning.Println(r.RemoteAddr, "GET /map/"+ds+" [401]")
+		err := fmt.Errorf("Unauthorized")
+		http.Error(w, err.Error(), http.StatusUnauthorized)
 	} else {
 		map_tmpl := "./templates/map_standard.html"
 		tmpl, _ := template.ParseFiles(map_tmpl)