+++ src/gospatial/app/logging.go
@@ -2,7 +2,7 @@ package app
 
 import (
 	"io"
-	"io/ioutil"
+	// "io/ioutil"
 	"log"
 	"os"
 )
@@ -18,12 +18,11 @@ var (
 )
 
 func init() {
-	Trace = log.New(ioutil.Discard, "TRACE: ", log.Ldate|log.Ltime|log.Lshortfile)
-	Info = log.New(os.Stdout, "INFO: ", log.Ldate|log.Ltime|log.Lshortfile)
-	Debug = log.New(os.Stdout, "DEBUG: ", log.Ldate|log.Ltime|log.Lshortfile)
-	Warning = log.New(os.Stdout, "WARNING: ", log.Ldate|log.Ltime|log.Lshortfile)
-	Error = log.New(os.Stderr, "ERROR: ", log.Ldate|log.Ltime|log.Lshortfile)
-	WebClient = log.New(os.Stdout, "CLIENT: ", log.Ldate|log.Ltime|log.Lshortfile)
+	Trace = log.New(os.Stdout, "[FIND] TRACE | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+	Info = log.New(os.Stdout, "[FIND] INFO  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+	Debug = log.New(os.Stdout, "[FIND] DEBUG | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+	Warning = log.New(os.Stdout, "[FIND] WARN  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+	Error = log.New(os.Stderr, "[FIND] ERROR | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 }
 
 func DebugMode(status bool) {
@@ -34,24 +33,16 @@ func DebugMode(status bool) {
 			Error.Fatal("Error opening file: %v", err)
 		}
 		// defer DebugModeLogFile.Close()
-		Trace = log.New(DebugModeLogFile, "[FIND] TRACE | ", log.Ldate|log.Ltime|log.Lshortfile)
-		Info = log.New(DebugModeLogFile, "[FIND] INFO  | ", log.Ldate|log.Ltime|log.Lshortfile)
-		Debug = log.New(DebugModeLogFile, "[FIND] DEBUG | ", log.Ldate|log.Ltime|log.Lshortfile)
-		Warning = log.New(DebugModeLogFile, "[FIND] WARN  | ", log.Ldate|log.Ltime|log.Lshortfile)
-		Error = log.New(DebugModeLogFile, "[FIND] ERROR | ", log.Ldate|log.Ltime|log.Lshortfile)
+		Trace = log.New(DebugModeLogFile, "[FIND] TRACE | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+		Info = log.New(DebugModeLogFile, "[FIND] INFO  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+		Debug = log.New(DebugModeLogFile, "[FIND] DEBUG | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+		Warning = log.New(DebugModeLogFile, "[FIND] WARN  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+		Error = log.New(DebugModeLogFile, "[FIND] ERROR | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 	} else {
-
-		// Trace = log.New(os.Stdout, "[SERVER] TRACE : ", log.LUTC|log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile)
-		// Info = log.New(os.Stdout, "[SERVER] INFO : ", log.LUTC|log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile)
-		// Debug = log.New(os.Stdout, "[SERVER] DEBUG : ", log.LUTC|log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile)
-		// Warning = log.New(os.Stdout, "[SERVER] WARNING : ", log.LUTC|log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile)
-		// Error = log.New(os.Stderr, "[SERVER] ERROR : ", log.LUTC|log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile)
-
 		Trace = log.New(os.Stdout, "[FIND] TRACE | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 		Info = log.New(os.Stdout, "[FIND] INFO  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 		Debug = log.New(os.Stdout, "[FIND] DEBUG | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 		Warning = log.New(os.Stdout, "[FIND] WARN  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 		Error = log.New(os.Stderr, "[FIND] ERROR | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-		// CRITICAL
 	}
 }